se agrego la funcion validate
